pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Redacted Official Hub"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
]
